Module linera_alloy_provider::fillers

Transaction Fillers

Fillers decorate a Provider, filling transaction details before they are sent to the network. Fillers are used to set the nonce, gas price, gas limit, and other transaction details, and are called before any other layer.

Structs

ChainIdFiller

A TxFiller that populates the chain ID of a transaction.

FillProvider

A Provider that applies one or more TxFillers.

GasFiller

A TxFiller that populates gas related fields in transaction requests if unset.

JoinFill

A layer that can fill in a TransactionRequest with additional information by joining two TxFillers. This struct is itself a TxFiller, and can be nested to compose any number of fill layers.

NonceFiller

A TxFiller that fills nonces on transactions.

SignerFiller

A layer that signs transactions locally.

Enums

FillerControlFlow

The control flow for a filler.

Traits

TxFiller

A layer that can fill in a TransactionRequest with additional information.