Trait linera_alloy_provider::fillers::TxFiller

pub trait TxFiller<N: Network = Ethereum>: Clone + Send + Sync + Debug {
    type Fillable: Send + Sync + 'static;

    // Required methods
    fn status(&self, tx: &N::TransactionRequest) -> FillerControlFlow;
    fn prepare<P, T>(
        &self,
        provider: &P,
        tx: &N::TransactionRequest,
    ) -> impl Send + Future<Output = TransportResult<Self::Fillable>>
       where P: Provider<T, N>,
             T: Transport + Clone;
    fn fill(
        &self,
        fillable: Self::Fillable,
        tx: SendableTx<N>,
    ) -> impl Send + Future<Output = TransportResult<SendableTx<N>>>;

    // Provided methods
    fn join_with<T>(self, other: T) -> JoinFill<Self, T>
       where T: TxFiller<N> { ... }
    fn continue_filling(&self, tx: &SendableTx<N>) -> bool { ... }
    fn ready(&self, tx: &N::TransactionRequest) -> bool { ... }
    fn finished(&self, tx: &N::TransactionRequest) -> bool { ... }
    fn prepare_and_fill<P, T>(
        &self,
        provider: &P,
        tx: SendableTx<N>,
    ) -> impl Send + Future<Output = TransportResult<SendableTx<N>>>
       where P: Provider<T, N>,
             T: Transport + Clone { ... }
}

A layer that can fill in a TransactionRequest with additional information.

Lifecycle Notes

The FillerControlFlow determines the lifecycle of a filler. Fillers may be in one of three states:

Missing: The filler is missing a required property to fill in the transaction request. TxFiller::status should return FillerControlFlow::Missing. with a list of the missing properties.

Ready: The filler is ready to fill in the transaction request. TxFiller::status should return FillerControlFlow::Ready.

Finished: The filler has filled in all properties that it can fill. TxFiller::status should return FillerControlFlow::Finished.

Required Associated Types

type Fillable: Send + Sync + 'static

The properties that this filler retrieves from the RPC. to fill in the TransactionRequest.

Required Methods

fn status(&self, tx: &N::TransactionRequest) -> FillerControlFlow

Return a control-flow enum indicating whether the filler is ready to fill in the transaction request, or if it is missing required properties.

fn prepare<P, T>( &self, provider: &P, tx: &N::TransactionRequest, ) -> impl Send + Future<Output = TransportResult<Self::Fillable>>
where P: Provider<T, N>, T: Transport + Clone,

Prepares fillable properties, potentially by making an RPC request.

fn fill( &self, fillable: Self::Fillable, tx: SendableTx<N>, ) -> impl Send + Future<Output = TransportResult<SendableTx<N>>>

Fills in the transaction request with the fillable properties.

Provided Methods

fn join_with<T>(self, other: T) -> JoinFill<Self, T>
where T: TxFiller<N>,

Joins this filler with another filler to compose multiple fillers.

fn continue_filling(&self, tx: &SendableTx<N>) -> bool

Returns true if the filler is should continnue filling.

fn ready(&self, tx: &N::TransactionRequest) -> bool

Returns true if the filler is ready to fill in the transaction request.

fn finished(&self, tx: &N::TransactionRequest) -> bool

Returns true if the filler is finished filling in the transaction request.

fn prepare_and_fill<P, T>( &self, provider: &P, tx: SendableTx<N>, ) -> impl Send + Future<Output = TransportResult<SendableTx<N>>>
where P: Provider<T, N>, T: Transport + Clone,

Prepares and fills the transaction request with the fillable properties.

Object Safety

This trait is not object safe.