Struct linera_alloy_provider::layers::AnvilProvider

pub struct AnvilProvider<P, T> { /* private fields */ }

A provider that wraps an AnvilInstance, preventing the instance from being dropped while the provider is in use.

Implementations

impl<P, T> AnvilProvider<P, T>
where P: Provider<T>, T: Transport + Clone,

pub fn new(inner: P, _anvil: Arc<AnvilInstance>) -> Self

Creates a new AnvilProvider with the given inner provider and anvil instance.