Struct linera_alloy_provider::ProviderBuilder

pub struct ProviderBuilder<L, F, N = Ethereum> { /* private fields */ }

A builder for constructing a Provider from various layers.

This type is similar to tower::ServiceBuilder, with extra complication around maintaining the network and transport types.

Implementations

impl ProviderBuilder<Identity, Identity, Ethereum>

pub const fn new() -> Self

Create a new ProviderBuilder.

impl<L, N> ProviderBuilder<L, Identity, N>

pub fn with_recommended_fillers( self, ) -> ProviderBuilder<L, JoinFill<JoinFill<JoinFill<Identity, GasFiller>, NonceFiller>, ChainIdFiller>, N>

Add preconfigured set of layers handling gas estimation, nonce management, and chain-id fetching.

pub fn with_gas_estimation( self, ) -> ProviderBuilder<L, JoinFill<Identity, GasFiller>, N>

Add gas estimation to the stack being built.

See GasFiller

pub fn with_nonce_management( self, ) -> ProviderBuilder<L, JoinFill<Identity, NonceFiller>, N>

Add nonce management to the stack being built.

See NonceFiller

pub fn fetch_chain_id( self, ) -> ProviderBuilder<L, JoinFill<Identity, ChainIdFiller>, N>

Add a chain ID filler to the stack being built. The filler will attempt to fetch the chain ID from the provider using Provider::get_chain_id. the first time a transaction is prepared, and will cache it for future transactions.

pub fn with_chain_id( self, chain_id: u64, ) -> ProviderBuilder<L, JoinFill<Identity, ChainIdFiller>, N>

Add a specific chain ID to the stack being built. The filler will fill transactions with the provided chain ID, regardless of the chain ID that the provider reports via Provider::get_chain_id.

impl<L, F, N> ProviderBuilder<L, F, N>

pub fn layer<Inner>( self, layer: Inner, ) -> ProviderBuilder<Stack<Inner, L>, F, N>

Add a layer to the stack being built. This is similar to tower::ServiceBuilder::layer.

Note:

Layers are added in outer-to-inner order, as in tower::ServiceBuilder. The first layer added will be the first to see the request.

pub fn filler<F2>(self, filler: F2) -> ProviderBuilder<L, JoinFill<F, F2>, N>

Add a transaction filler to the stack being built. Transaction fillers are used to fill in missing fields on transactions before they are sent, and are all joined to form the outermost layer of the stack.

pub fn signer<S>( self, signer: S, ) -> ProviderBuilder<L, JoinFill<F, SignerFiller<S>>, N>

Add a signer layer to the stack being built.

See SignerFiller.

pub fn network<Net: Network>(self) -> ProviderBuilder<L, F, Net>

Change the network.

By default, the network is Ethereum. This method must be called to configure a different network.

builder.network::<Arbitrum>()

pub fn with_chain( self, chain: NamedChain, ) -> ProviderBuilder<Stack<ChainLayer, L>, F, N>

Add a chain layer to the stack being built. The layer will set the client’s poll interval based on the average block time for this chain.

Does nothing to the client with a local transport.

pub fn on_provider<P, T>(self, provider: P) -> F::Provider
where L: ProviderLayer<P, T, N>, F: TxFiller<N> + ProviderLayer<L::Provider, T, N>, P: Provider<T, N>, T: Transport + Clone, N: Network,

Finish the layer stack by providing a root Provider, outputting the final Provider type with all stack components.

pub fn on_client<T>(self, client: RpcClient<T>) -> F::Provider
where L: ProviderLayer<RootProvider<T, N>, T, N>, F: TxFiller<N> + ProviderLayer<L::Provider, T, N>, T: Transport + Clone, N: Network,

Finish the layer stack by providing a root RpcClient, outputting the final Provider type with all stack components.

This is a convenience function for ProviderBuilder::provider<RpcClient>.

pub async fn on_builtin(self, s: &str) -> Result<F::Provider, TransportError>
where L: ProviderLayer<RootProvider<BoxTransport, N>, BoxTransport, N>, F: TxFiller<N> + ProviderLayer<L::Provider, BoxTransport, N>, N: Network,

Finish the layer stack by providing a connection string for a built-in transport type, outputting the final Provider type with all stack components.

pub fn on_http(self, url: Url) -> F::Provider
where L: ProviderLayer<ReqwestProvider<N>, Http<Client>, N>, F: TxFiller<N> + ProviderLayer<L::Provider, Http<Client>, N>, N: Network,

Build this provider with an Reqwest HTTP transport.

impl<L, F> ProviderBuilder<L, F, Ethereum>

pub fn on_anvil(self) -> F::Provider
where F: TxFiller<Ethereum> + ProviderLayer<L::Provider, Http<Client>, Ethereum>, L: ProviderLayer<AnvilProvider<RootProvider<Http<Client>>, Http<Client>>, Http<Client>>,

Build this provider with anvil, using an Reqwest HTTP transport.

pub fn on_anvil_with_signer( self, ) -> <JoinFill<F, SignerFiller<EthereumSigner>> as ProviderLayer<L::Provider, Http<Client>>>::Provider
where F: TxFiller<Ethereum> + ProviderLayer<L::Provider, Http<Client>, Ethereum>, L: ProviderLayer<AnvilProvider<RootProvider<Http<Client>>, Http<Client>>, Http<Client>>,

Build this provider with anvil, using an Reqwest HTTP transport. This function configures a signer backed by anvil keys, and is intended for use in tests.

pub fn on_anvil_with_config(self, f: impl FnOnce(Anvil) -> Anvil) -> F::Provider
where F: TxFiller<Ethereum> + ProviderLayer<L::Provider, Http<Client>, Ethereum>, L: ProviderLayer<AnvilProvider<RootProvider<Http<Client>>, Http<Client>>, Http<Client>>,

Build this provider with anvil, using an Reqwest HTTP transport. The given function is used to configure the anvil instance.

pub fn on_anvil_with_signer_and_config( self, f: impl FnOnce(Anvil) -> Anvil, ) -> <JoinFill<F, SignerFiller<EthereumSigner>> as ProviderLayer<L::Provider, Http<Client>>>::Provider
where F: TxFiller<Ethereum> + ProviderLayer<L::Provider, Http<Client>, Ethereum>, L: ProviderLayer<AnvilProvider<RootProvider<Http<Client>>, Http<Client>>, Http<Client>>,

Build this provider with anvil, using an Reqwest HTTP transport. The given function is used to configure the anvil instance. This function configures a signer backed by anvil keys, and is intended for use in tests.