Trait linera_alloy_provider::WalletProvider

pub trait WalletProvider<N: Network = Ethereum> {
    type Signer: NetworkSigner<N>;

    // Required methods
    fn signer(&self) -> &Self::Signer;
    fn signer_mut(&mut self) -> &mut Self::Signer;

    // Provided methods
    fn default_signer_address(&self) -> Address { ... }
    fn has_signer_for(&self, address: &Address) -> bool { ... }
    fn signer_addresses(&self) -> impl Iterator<Item = Address> { ... }
}

Trait for Providers, Fill stacks, etc, which contain NetworkSigner.

Required Associated Types

type Signer: NetworkSigner<N>

The underlying NetworkSigner type contained in this stack.

Required Methods

fn signer(&self) -> &Self::Signer

Get a reference to the underlying signer.

fn signer_mut(&mut self) -> &mut Self::Signer

Get a mutable reference to the underlying signer.

Provided Methods

fn default_signer_address(&self) -> Address

Get the default signer address.

fn has_signer_for(&self, address: &Address) -> bool

Check if the signer can sign for the given address.

fn signer_addresses(&self) -> impl Iterator<Item = Address>

Get an iterator of all signer addresses. Note that because the signer always has at least one address, this iterator will always have at least one element.

Object Safety

This trait is not object safe.